pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-cookie-vault"
version = "1.0.0"
description = "Secure cookie and session storage for MCP (Model Context Protocol)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MCP Cookie Vault Contributors"}
]
keywords = [
    "mcp",
    "cookie",
    "session",
    "vault",
    "playwright",
    "automation",
    "browser-automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "mcp>=1.0.0",
    "cryptography>=41.0.0",
    "keyring>=24.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "playwright>=1.40.0",
]

[project.scripts]
mcp-cookie-vault = "mcp_cookie_vault.server:main"

[project.urls]
Homepage = "https://github.com/lakismak/mcp-cookie-vault"
Documentation = "https://github.com/lakismak/mcp-cookie-vault#readme"
Repository = "https://github.com/lakismak/mcp-cookie-vault"
Issues = "https://github.com/lakismak/mcp-cookie-vault/issues"

[tool.hatch.build.targets.wheel]
packages = ["mcp_cookie_vault"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["mcp_cookie_vault", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "S"]
ignore = ["S101"]  # Allow assert in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
